# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "帖子服务，提供帖子发布、查询、管理等功能。",
        "title": "Post Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
        "/api/v1/post/admin/posts": {
            "get": {
                "description": "出于管理目的，根据各种过滤条件检索分页的帖子列表。使用查询参数进行过滤和分页。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-posts (管理员-帖子)"
                ],
                "summary": "按条件列出帖子 (管理员)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint64",
                        "description": "按精确的帖子 ID 过滤",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按帖子标题过滤（模糊匹配）",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按作者用户名过滤（模糊匹配）",
                        "name": "author_username",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "按帖子状态过滤 (0=待审核, 1=已审核, 2=已拒绝)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "按官方标签过滤 (例如, 0=无, 1=官方认证)",
                        "name": "official_tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "按最小浏览量过滤",
                        "name": "view_count_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "按最大浏览量过滤",
                        "name": "view_count_max",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "default": "created_at",
                        "description": "排序字段 (created_at 或 updated_at)",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "是否降序排序 (true 为 DESC, false/省略为 ASC)",
                        "name": "order_desc",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int",
                        "description": "页码（从 1 开始）",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int",
                        "description": "每页帖子数量",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子检索成功\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.ListPostsAdminResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的输入参数（例如，无效的 page, page_size, status）\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "检索帖子时发生内部服务器错误\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
        "/api/v1/post/admin/posts/audit": {
            "post": {
                "description": "管理员更新帖子的状态（以及可选的原因）。需要在请求体中提供审核详情。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-posts (管理员-帖子)"
                ],
                "summary": "审核帖子",
                "parameters": [
                    {
                        "description": "审核帖子请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuditPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子审核成功\" // \u003c--- 修改 (无 Data)",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的请求负载（例如，缺少字段，无效的状态）\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "404": {
                        "description": "帖子未找到\" // \u003c-- 添加404情况",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "审核过程中发生内部服务器错误\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
        "/api/v1/post/admin/posts/{id}/official-tag": {
            "put": {
                "description": "管理员更新特定帖子的官方标签。需要在 URL 路径中提供帖子 ID，并在请求体中提供标签详情。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-posts (管理员-帖子)"
                ],
                "summary": "更新帖子官方标签 (管理员)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint64",
                        "description": "要更新的帖子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新官方标签请求体 (请求体中的 PostID 是冗余的，请使用路径中的 ID)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOfficialTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "官方标签更新成功\" // \u003c--- 修改 (无 Data)",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的请求负载，无效的标签值，或路径 ID 与请求体 ID 不匹配\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "404": {
                        "description": "帖子未找到\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "更新标签时发生内部服务器错误\" // \u003c--- 修改",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
        "/api/v1/post/admin/posts/{post_id}": {
            "delete": {
                "description": "管理员软删除指定ID的帖子 (Admin soft deletes a post with the specified ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员删除帖子 (Admin delete post)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID (Post ID)",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子删除成功",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的帖子ID格式",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "401": {
                        "description": "管理员未登录或无权限",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "404": {
                        "description": "帖子未找到",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "删除帖子时发生内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
...
        "/api/v1/post/posts": {
            "post": {
                "description": "使用提供的详情（作为独立表单字段）和图片文件创建一个新帖子。请求体应为 multipart/form-data。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts (帖子)"
                ],
                "summary": "创建新帖子 (独立表单字段及图片)",
                "parameters": [
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "帖子标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 1000,
                        "type": "string",
                        "description": "帖子内容",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "number",
                        "description": "单价 (可选, 大于等于0)",
                        "name": "price_per_unit",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "联系方式 (可选)",
                        "name": "contact_info",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "作者ID",
                        "name": "author_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "url",
                        "description": "作者头像 URL (可选, 需为有效URL)",
                        "name": "author_avatar",
                        "in": "formData"
                    },
                    {
                        "maxLength": 50,
                        "type": "string",
                        "description": "作者用户名",
                        "name": "author_username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "帖子图片文件 (可多选)",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子创建成功",
                        "schema": {
                            "$ref": "#/definitions/vo.PostDetailResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的请求负载或文件处理错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "创建帖子时发生内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
...
        "/api/v1/post/posts/{id}": {
            "delete": {
                "description": "通过帖子的 ID 软删除一个帖子。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts (帖子)"
                ],
                "summary": "删除指定ID的帖子",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint64",
                        "description": "帖子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子删除成功",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的帖子 ID 格式",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "删除帖子时发生内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        },
        "/api/v1/post/posts/{post_id}": {
            "get": {
                "description": "通过帖子的 ID 检索特定帖子的详细信息。同时，如果用户已登录（通过中间件注入UserID），则会尝试增加浏览量。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts (帖子)"
                ],
                "summary": "获取指定ID的帖子详情 (公开)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint64",
                        "description": "帖子 ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户 ID (由网关/中间件注入)",
                        "name": "X-User-ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子详情检索成功",
                        "schema": {
                            "$ref": "#/definitions/vo.PostDetailResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "无效的帖子 ID 格式",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    },
                    "500": {
                        "description": "检索帖子详情时发生内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/vo.BaseResponseWrapper"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AuditPostRequest": {
            "type": "object",
            "required": [
                "post_id"
            ],
            "properties": {
                "post_id": {
                    "description": "为 PostID 也添加一个 example",
                    "type": "integer",
                    "example": 123
                },
                "reason": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "内容符合规范"
                },
                "status": {
                    "description": "Status 表示帖子的审核状态。\n0: 待审核 (Pending)\n1: 审核通过 (Approved)\n2: 拒绝 (Rejected)",
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 0
                }
            }
        },
        "dto.UpdateOfficialTagRequest": {
            "type": "object",
            "required": [
                "official_tag",
                "post_id"
            ],
            "properties": {
                "official_tag": {
                    "description": "新的官方标签值，必填，并限制范围 (假设最大值为 3)",
                    "type": "integer",
                    "maximum": 3,
                    "minimum": 0
                },
                "post_id": {
                    "description": "帖子ID，必填",
                    "type": "integer"
                }
            }
        },
        "enums.OfficialTag": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-comments": {
                "OfficialTagCertified": "官方认证",
                "OfficialTagDeposit": "预付保证金",
                "OfficialTagNone": "无标签",
                "OfficialTagRapid": "急速响应"
            },
            "x-enum-varnames": [
                "OfficialTagNone",
                "OfficialTagCertified",
                "OfficialTagDeposit",
                "OfficialTagRapid"
            ]
        },
...
        "vo.BaseResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "成功时为 0, 错误时为具体错误码",
                    "type": "integer",
                    "example": 0
                },
                "message": {
                    "description": "成功或错误消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "vo.ListPostsAdminByConditionResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "description": "帖子列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.PostResponse"
                    }
                },
                "total": {
                    "description": "帖子总数",
                    "type": "integer"
                }
            }
        },
        "vo.ListPostsAdminResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "使用具体的 vo.ListPostsAdminByConditionResponse",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.ListPostsAdminByConditionResponse"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "vo.PostDetailResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/vo.PostDetailVO"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "vo.PostDetailVO": {
            "type": "object",
            "properties": {
                "author_avatar": {
                    "description": "作者头像URL",
                    "type": "string"
                },
                "author_id": {
                    "description": "作者ID",
                    "type": "string"
                },
                "author_username": {
                    "description": "作者用户名",
                    "type": "string"
                },
                "contact_info": {
                    "description": "联系方式 (手机号、微信号、QQ号等)",
                    "type": "string"
                },
                "content": {
                    "description": "--- 来自 PostDetail 实体 ---",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "--- 来自 Post 实体 ---",
                    "type": "integer"
                },
                "images": {
                    "description": "--- 来自 PostDetailImage 实体列表 ---\nImages 字段存储了帖子的所有详情图片，并已按 DisplayOrder 排序。",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.PostImageVO"
                    }
                },
                "official_tag": {
                    "description": "官方标签 (参考 enums.OfficialTag)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.OfficialTag"
                        }
                    ]
                },
                "price_per_unit": {
                    "description": "单价 (单位：元)",
                    "type": "number"
                },
                "title": {
                    "description": "帖子标题",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "view_count": {
                    "description": "浏览量",
                    "type": "integer"
                }
            }
        },
        "vo.PostImageVO": {
            "type": "object",
            "properties": {
                "display_order": {
                    "description": "图片展示顺序",
                    "type": "integer"
                },
                "image_url": {
                    "description": "图片URL",
                    "type": "string"
                },
                "object_key": {
                    "description": "图片在COS中的ObjectKey",
                    "type": "string"
                }
            }
        },
...
    }
}